+++ src/Types.hpp
@@ -254,7 +254,7 @@ struct MGLTexture {
 	int max_level;
 
 	int compare_func;
-	int anisotropy;
+	float anisotropy;
 
 	bool depth;
